Upload photoecc_mod code as of April 5th, 2021
This is the complete code for my photo-eccentric model, including data download, light curve processing, transit modeling, eccentricity re-sampling, plotting both input data and results, and storing the results. While the code works well for its purpose, it still needs significant overhaul before making it accessible to others. This version is a single-system example version for HIP-97166 with 1 transiting planet at 10.3 days. The impact parameter for this planet is poorly fit and leads to a significant overestimation of eccentricity. Working on figuring out what leads to this discrepancy and whether or not it can be mitigated in a self-contained way. When the code and data are downloaded, the paths within the code will need to be changed to fit the user's directory setup. I'm still learning more about true python software development, but hopefully I will have a more professional set-up soon!